refactored firebase auth class
